+++ src/main/java/wooteco/subway/SpecificExceptionHandler.java
@@ -1,21 +1,19 @@
 package wooteco.subway;
 
-import org.springframework.dao.DataAccessException;
+import org.springframework.core.Ordered;
+import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import wooteco.subway.exception.repository.DataNotFoundException;
 import wooteco.subway.exception.repository.DuplicatedFieldException;
-import wooteco.subway.exception.repository.RepositoryException;
+import wooteco.subway.exception.service.ObjectNotFoundException;
+import wooteco.subway.exception.service.ValidationFailureException;
 
+@Order(Ordered.HIGHEST_PRECEDENCE)
 @RestControllerAdvice
-public class RepositoryExceptionHandler {
-
-    @ExceptionHandler({RepositoryException.class, DataAccessException.class})
-    public ResponseEntity<String> handleRepositoryException(final RuntimeException e) {
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
-    }
+public class SpecificExceptionHandler {
 
     @ExceptionHandler(DataNotFoundException.class)
     public ResponseEntity<String> handleDataNotFoundException(final DataNotFoundException e) {
@@ -24,6 +22,16 @@ public class RepositoryExceptionHandler {
 
     @ExceptionHandler(DuplicatedFieldException.class)
     public ResponseEntity<String> handleDuplicatedFieldException(final DuplicatedFieldException e) {
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
+    }
+
+    @ExceptionHandler(ObjectNotFoundException.class)
+    public ResponseEntity<String> handleObjectNotFoundException(final ObjectNotFoundException e) {
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
+    }
+
+    @ExceptionHandler(ValidationFailureException.class)
+    public ResponseEntity<String> handleValidationFailureException(final ValidationFailureException e) {
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
     }
 }
